corrections
Hi Attah, sorry for my late Answer. I read your E-Mail late because it was not my main mail Address. correction of capitalization and some spelling mistakes. I have written many of the words in lower case because they are not nouns in German.  But you are right, at the beginning of a sentence capitalize. 

The terms I did not know the translation and found nothing about that are: 
- Line 753: bi-level. I found "Doppelstock" or "zweistufige verarbeitung" but i don´t think that is correct for that.
- Line 593: pending-held. I found "Ausstehend gehalten" but it sounds crazy for me.

- Line 651: "...Druckern und CUBS-Instanzen..." I found no other Translation for CUPS-instances so I think it is correct. I correct it to "...Druckern und CUPS-Instanzen...".